move `kg` to tdc/utils folder
