+++ tdc/__init__.py
@@ -2,4 +2,4 @@ from .evaluator import Evaluator
 from .oracles import Oracle
 from .benchmark_deprecated import BenchmarkGroup
 from .tdc_hf import tdc_hf_interface
-from .knowledge_graph import KnowledgeGraph
+from tdc.utils.knowledge_graph import KnowledgeGraph
